Add menu support for any model
Example app now illustrates menu functionality with built-in page
schema. Pages include a 'path' field which defines a relative path to
view the page. The menu link correctly handles this path, but Macondo
needs to also support serving that path by rendering it through the
template specified in the page's 'template' field.